backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==2.0.1
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.66.0
plotly==5.24.1